+++ modules/display_format/builtin/include/dispBuiltin.hpp
@@ -12,10 +12,17 @@
 #include "ArrayOf.hpp"
 #include "Evaluator.hpp"
 //=============================================================================
-namespace Nelson::FunctionsGateway {
+namespace Nelson {
 //=============================================================================
-ArrayOfVector
-addpathBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+namespace DisplayFormatGateway {
+    //=============================================================================
+    ArrayOfVector
+    dispBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+    //=============================================================================
+    ArrayOfVector
+    generic_dispBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+    //=============================================================================
+}
 //=============================================================================
 } // namespace Nelson
 //=============================================================================