+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## UNRELEASED
+
+- plug `display_format` module (format, disp, display, echo, formattedDisplayText)
+
+- builtin loaded by gateway.
+
 ## 0.1.0
 
 ### Added